実装：全テーブルマイグレーション及びリレーション作成　修正 テストケース作成

+++ database/migrations/2017_08_04_054425_create_dailies_table.php
@@ -22,7 +22,7 @@ class CreateDailiesTable extends Migration
             $table->date('date');
             $table->time('start');
             $table->time('end');
-            $table->double('rest', 2, 1);
+            $table->integer('rest');
             $table->double('time', 2, 1);
             $table->integer('cost');
             $table->text('note');
